+++ entrypoint.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
+set -e
 
 /setuid.sh \
-&& su -m dashcam /viofosync.sh \
-&& [[ -z $RUN_ONCE ]] \
-&& crond -f
+&& su -m dashcam /viofosync.sh
+
+if [[ -z $RUN_ONCE ]]; then
+    crond -f
+fi